+++ frontend/src/components/ui/Graph/CircleDaySelectButton.tsx
@@ -7,8 +7,7 @@ import ButtonNext from '@common/ButtonNext';
 import { ENGLISH_DAYS, ENGLISH_DAYS_TO_KOREAN_DAYS } from '@constants/congestion';
 
 import { GraphContext } from '.';
-
-import type { EnglishDaysType } from 'types';
+import type { EnglishDaysType } from '../../../types/congestion';
 
 const isEnglishDays = (day: string): day is EnglishDaysType => {
   return ENGLISH_DAYS.includes(day as EnglishDaysType);
