+++ frontend/src/hooks/tanstack-query/station-details/useStationCongestionStatistics.ts
@@ -8,7 +8,7 @@ import { serverStore } from '@stores/serverStore';
 import { SERVERS } from '@constants';
 import { ERROR_MESSAGES } from '@constants/errorMessages';
 
-import type { CongestionStatistics } from '../../../types';
+import type { CongestionStatistics } from '../../../types/congestion';
 
 export const fetchStationDetails = async (selectedStationId: number) => {
   const mode = serverStore.getState();
